Added capability to detect e-mail addresses that shall be used for tests with notification mails.
Notification mails do not contain a personal prefix like 'John Doe' <jd@ox.invalid> but only the plain address <jd@ox.invalid>. Recipients or senders that shall be used for notification tests can be marked like 'to: users:users_a (notification)' at the corresponding e-mail fixtures. The ' (notification') part is detected from the fixtures and the personal prefix (displayname) is not added for these entries. This detection also works when defining multiple recipients for an e-mail.