Fixed: Bug 28004 - Using the App suite while updating frontend packages makes the ui unuseable
Concatenated JS modules always return HTTP status 200. Errors are
returned as short JS snippets which call the appropriate AIPs in the
client.

Until now, the "Expires" header was always in the future. Now, the data
is buffered until it is known if there are errors or not. Then, the
"Expires" header is set to "0" if there are errors, or left as before
ca. one year in the future, if there are no errors.