Started with refactoring of facets.
The goal is to provide a cleaner API for clients
and developers that build new ModuleSearchDrivers.
Therefore some logic has been moved out of the
FacetType interface into Facet itself. Facet is
an interface now and there are three styles of
facets that are realized as subclasses:

1. SimpleFacet ('style':'simple')
   The equivalent to former type FieldFacet.
2. DefaultFacet ('style':'default')
   The equivalent to former type Facet.
3. ExclusiveFacet ('style':'exclusive')
   This one is new and is used for facets with
   values that are mutually exclusive.

Clients can distinguish facets by their 'style'
attribute now and decide to provide a better
user experience by handling the respective
quirks in different ways. A FacetVisitor has
been introduced for an easy handling of facet
collections. This reduces decision trees that
formerly operated on FacetType flags.

BasicMailDriver shows how the new classes are
meant to be used in search drivers. It can be
seen as best practice to create a subclass per
FacetType in every module that encapsulates
boilerplate code and allows for speaking names
instead of generic constructors with long para-
meter lists.