Bug #43941: Changed c.o.file.storage.json.actions.accounts.AllAction to return all filestorage accounts and exceptions.
Exceptions will be appended to the filestorageaccounts where they occur.
Accounts with errors will be marked with 'hasError':'true'

Introduced the interface LoginAwareFileStorageServiceExtension to test
the connection of accounts which does not depend on oauth accounts. The
connections of this accounts will be tested when they are created or
updated.