Bug #44836: Extended deleteuser clt with reassign and no-reassign parameter
-reassign allows the admin to assign all shared data to the specified
user. All files will be moved to the filestore of the given user. If
omitted all data will be assigned to ctx admin and all data will be
moved to the ctx filestore.

-no-reassign allows the admin to remove all shared data instead. Shared
file and folders (and all its content) will be deleted. Tasks and
Appointment are still assigned to the ctx admin.