Fix for bug 67667: Avoid resetting PartStat
Avoid to reset the participant status to NEEDS_ACTION in case a Mac Calendar sets this value, even if the status has already been set.
Mac calendar seems to try enforcing a reset status for special action like e.g. when the client thinks rescheduling has been performed, even if only an exception has been added.
Also aligns the behavior with a normal client, which does not offer to set the status to NEEDS_ACTION.