MWB-357: Don't close client implicit
The service does only return a HttpClient, not a ClosableHttpClient. Therefore the method close() can't be called for the managed client. This means the ClosableListener was only executed when the HTTP client was closed via the service. Therefore the "hack" was superfluous by now.
When the client is closed now, it will be removed explicitly from the cache.