MWB-366: Remove GROUP BY clause
The primary key of the table oxfolder_tree consists of the context ID and the folder ID. So when adding the CID in the WHERE clause of a query targeting this table, uniqueness of the folder ID is guaranteed. Thus it doesn't make sense to use the GROUP BY clause on the folder ID. Therefore removed the clause.