MWB-371: Use preemtive auth for sipgate
sipgate uses a RESTful API which uses basic auth to authenticate the request
HTTP client seems to struggle because of the missing `WWW-Authenticate` header on the returning 401-response when trying "normal" BasicAuth, which will always try to get a valid response without credentials. Therefore a re-try is not made and the request fails.