MWB-375: Only send mail to the originator
Sending mails to all attendees of the event doesn't make sense in cases of REFRESH or DECLINECOUNTER. Instead only the originator of the request should be informed. Therefore only send one mail to the originator for certain iTIP methods.
Because the originator can't be determined from the calendar event, added methods that surely gets the originator based on the mails from header.