Generate diffs for unset properties, too
In case a complex property was removed from the updated event, the expression in `considerCollectionUpdate()` failed. Thus e.g. attachments could have been removed and the EventUpdated said that there are no changes (`isEmpty() == true`). This caused problems in iTIP where the changes or rather the removal could not be applied.