MW-1378: Further hardening
Localized error messages
Removed unused code
Invalidate client if credentials are outdated. Note: Didn't add the password to the cache. If password is changed, the cache will have a miss and thus create a new client. The old client with outdated credentials will then only be invalidated by the one day timeout. With the current approach we properly close the outdated client.