Adds the possibility to enable/disable packages via the OX_BLACKLISTED_PACKAGES environment variable. The value of the variable is computed according to the given values in the features.status and packages.status subtrees:
When a packages is deemed to be disabled it is added to the list. A package is disabled, if:

It is listed under packages.status with a value of 'disabled':

  packages:
    status:
      - open-xchange-authentication-database: 'disabled'

OR if it is disabled as part of a feature in features.status:

  features:
    status:
      - admin: 'disabled'

This would disable all packages that are listed at features.definitions.admin, UNLESS they are enabled again in the
packages section. A package or feature can be enabled in their respective sections by supplying a value of 'enabled'

See the values.yaml file for default values